// src/app/functionnal-views/travel-management/interfaces/travel.ts
import { TravelState } from "../enums/travel-state";
import { TravelType } from "../enums/travel-type";

export interface ITravel {
    id: string;
    departAgencyId: string;
    arrivalAgencyId: string;
    travelPathId: string;
    busId: string;
    busDriverId: string;
    plannedDepartDatetime: Date;
    effectiveDepartDatetime?: Date;
    arrivalDatetime?: Date;
    travelType: TravelType;
    travelState: TravelState;
    ticketPrice: number;
    reservationFees: number;
}